퍼미션 체크하는 부분 수정

+++ app/src/main/java/kr/co/valuesys/vlog/mobile/activity/SplashActivity.java
@@ -1,17 +1,21 @@
 package kr.co.valuesys.vlog.mobile.activity;
 
+import android.content.ActivityNotFoundException;
 import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.Signature;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.databinding.DataBindingUtil;
 
+import android.net.Uri;
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.provider.Settings;
 import android.text.TextUtils;
 import android.util.Base64;
 
@@ -20,6 +24,7 @@ import com.facebook.login.LoginManager;
 import com.kakao.auth.ISessionCallback;
 import com.kakao.auth.Session;
 
+import java.io.File;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
@@ -28,6 +33,7 @@ import kr.co.valuesys.vlog.mobile.common.FileManager;
 import kr.co.valuesys.vlog.mobile.common.LogUtil;
 import kr.co.valuesys.vlog.mobile.R;
 import kr.co.valuesys.vlog.mobile.common.PermissionUtils;
+import kr.co.valuesys.vlog.mobile.common.SimpleAlert;
 import kr.co.valuesys.vlog.mobile.databinding.ActivitySplashBinding;
 
 import static kr.co.valuesys.vlog.mobile.common.Constants.FaceBook;
@@ -45,6 +51,8 @@ public class SplashActivity extends AppCompatActivity {
 
     private PermissionUtils m_permissionUtils;
 
+    private AlertDialog alertDialog;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -52,7 +60,10 @@ public class SplashActivity extends AppCompatActivity {
         binding = DataBindingUtil.setContentView(this, R.layout.activity_splash);
 
 //        getHashKey();
-        FileManager.deleteTempFiles(this);
+        if (FileManager.checkStoragePermission(this)) {
+            FileManager.deleteTempFiles(this);
+        }
+
 
         binding.spLogo.postDelayed(() -> {
 
@@ -140,46 +151,67 @@ public class SplashActivity extends AppCompatActivity {
 
     }
 
-    private void presentNext() {
+//    private void presentNext() {
+//
+//        binding.spLogo.postDelayed(() -> {
+//
+//            if (!TextUtils.isEmpty(session)) {
+//
+//                if (m_permissionUtils.checkPermission() == false) {
+//                    m_permissionUtils.requestPermission();
+//
+//                } else {
+//
+//                    MobileApplication.getContext().writeLog(MobileApplication.getContext().getLoginkName());
+//
+//                    Intent intent = new Intent(SplashActivity.this, MainActivity.class);
+//                    startActivity(intent);
+//                    finish();
+//                }
+//
+//
+//            } else {
+//
+//                Intent intent = new Intent(this, LoginActivity.class);
+//                startActivity(intent);
+//                finish();
+//
+//            }
+//
+//        }, 1000);
+//
+//    }
 
-        binding.spLogo.postDelayed(() -> {
-
-            if (!TextUtils.isEmpty(session)) {
-
-                if (m_permissionUtils.checkPermission() == false) {
-                    m_permissionUtils.requestPermission();
-
-                } else {
-
-                    MobileApplication.getContext().writeLog(MobileApplication.getContext().getLoginkName());
-
-                    Intent intent = new Intent(SplashActivity.this, MainActivity.class);
-                    startActivity(intent);
-                    finish();
-                }
-
-
-            } else {
-
-                Intent intent = new Intent(this, LoginActivity.class);
-                startActivity(intent);
-                finish();
-
-            }
-
-        }, 1000);
-
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-//        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (m_permissionUtils.permissionResult(requestCode, permissions, grantResults) == false) {
-            m_permissionUtils.requestPermission();
-        } else {
-            presentNext();
-        }
-    }
+//    @Override
+//    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+////        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+//        if (m_permissionUtils.permissionResult(requestCode, permissions, grantResults) == false) {
+//            m_permissionUtils.requestPermission();
+//
+//            alertDialog = SimpleAlert.createAlert(this, "권한을 허용해 주세요", false, dialog -> {
+//
+//                dialog.dismiss();
+//
+//                try {
+//                    Intent intent = new Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
+//                            .setData(Uri.parse("package:" + getPackageName()));
+//                    startActivity(intent);
+//                    finish();
+//                } catch (ActivityNotFoundException e) {
+//                    e.printStackTrace();
+//
+//                    Intent intent = new Intent(Settings.ACTION_MANAGE_APPLICATIONS_SETTINGS);
+//                    startActivity(intent);
+//                    finish();
+//                }
+//
+//            });
+//            alertDialog.show();
+//
+//        } else {
+//            presentNext();
+//        }
+//    }
 
     /**
      * 카카오 sdk 홈페이지에 입력 해야할 해시키값 , 페이스북도 마찬가지
